feat(migration backend) : modifications des requêtes gift - fin de migration

+++ backend/app/schemas/gift/gift_base.py
@@ -12,13 +12,13 @@ class GiftBase(CamelModel):
     description: Optional[str] = None
     url: Optional[str] = None
     quantite: int = 1
-    prix: Optional[Decimal] = None
+    prix: Optional[float] = None
     commentaire: Optional[str] = None
     priorite: int
     statut: GiftStatusEnum = GiftStatusEnum.DISPONIBLE  # valeur par défaut
     marque: Optional[str] = None
     magasin: Optional[str] = None
-    prixReel: Optional[Decimal] = None
-    fraisPort: Optional[Decimal] = None
+    prixReel: Optional[float] = None
+    fraisPort: Optional[float] = None
